server/users/serializers.py
from rest_framework.authtoken.admin import User

from rest_framework import serializers


class UserInfoSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ('id', 'username', 'email')