server/requirements.txt
asgiref==3.8.1
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
python-http-client==3.3.7
sendgrid==3.6.5
sendgrid-django==4.2.0
six==1.17.0
sqlparse==0.5.3
redis~=6.2.0